Phase 3: extract lead validation, rate limit submitLead, harden import
- Move validateLeadInput, normalizeBudget, and constants to convex/lib/leadValidation.ts
- Add by_email index and block submitLead when 3+ leads from same email in the last hour
- Import migration validates each row; return inserted, skipped, and invalid counts
- Document supabaseId dedup and invalid-row handling in convex-migration-notes
