+++ convex/_generated/api.d.ts
@@ -9,6 +9,7 @@
  */
 
 import type * as leads from "../leads.js";
+import type * as lib_leadValidation from "../lib/leadValidation.js";
 import type * as migrations_importLeads from "../migrations/importLeads.js";
 
 import type {
@@ -19,6 +20,7 @@ import type {
 
 declare const fullApi: ApiFromModules<{
   leads: typeof leads;
+  "lib/leadValidation": typeof lib_leadValidation;
   "migrations/importLeads": typeof migrations_importLeads;
 }>;
