+++ convex/leads.ts
@@ -1,55 +1,10 @@
 import { ConvexError, v } from "convex/values";
 import { internal } from "./_generated/api";
 import { internalAction, mutation } from "./_generated/server";
+import { validateLeadInput } from "./lib/leadValidation";
 
-const BUDGET_OPTIONS = ["< €5k", "€5k-€10k", "€10k-€25k", "€25k+"] as const;
-const MAX_NAME_LENGTH = 200;
-const MAX_DESCRIPTION_LENGTH = 5000;
-const EMAIL_REGEX = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
-
-function normalizeBudget(budget: string | undefined) {
-  const trimmed = budget?.trim();
-  if (!trimmed) {
-    return undefined;
-  }
-  if (!BUDGET_OPTIONS.includes(trimmed as (typeof BUDGET_OPTIONS)[number])) {
-    throw new ConvexError("Invalid budget range.");
-  }
-  return trimmed;
-}
-
-function validateLeadInput(args: {
-  name: string;
-  email: string;
-  budget?: string;
-  description: string;
-}) {
-  const name = args.name.trim();
-  const email = args.email.trim();
-  const description = args.description.trim();
-  const budget = normalizeBudget(args.budget);
-
-  if (!name) {
-    throw new ConvexError("Name is required.");
-  }
-  if (name.length > MAX_NAME_LENGTH) {
-    throw new ConvexError("Name is too long.");
-  }
-  if (!email) {
-    throw new ConvexError("Email is required.");
-  }
-  if (!EMAIL_REGEX.test(email)) {
-    throw new ConvexError("Invalid email address.");
-  }
-  if (!description) {
-    throw new ConvexError("Project description is required.");
-  }
-  if (description.length > MAX_DESCRIPTION_LENGTH) {
-    throw new ConvexError("Project description is too long.");
-  }
-
-  return { name, email, description, budget };
-}
+const RATE_LIMIT_WINDOW_MS = 3_600_000;
+const RATE_LIMIT_MAX_SUBMISSIONS = 3;
 
 export const submitLead = mutation({
   args: {
@@ -61,6 +16,17 @@ export const submitLead = mutation({
   returns: v.object({ success: v.boolean() }),
   handler: async (ctx, args) => {
     const lead = validateLeadInput(args);
+
+    const cutoff = Date.now() - RATE_LIMIT_WINDOW_MS;
+    const recent = await ctx.db
+      .query("leads")
+      .withIndex("by_email", (q) => q.eq("email", lead.email))
+      .filter((q) => q.gt(q.field("createdAt"), cutoff))
+      .collect();
+
+    if (recent.length >= RATE_LIMIT_MAX_SUBMISSIONS) {
+      throw new ConvexError("Please wait before submitting again.");
+    }
 
     const leadId = await ctx.db.insert("leads", {
       ...lead,
